IPipingProvider を廃止し、 IPipingStore がそれを受け持つ様に修正。 また、 API を Pipelines.Pipe に近い感じに調整
