+++ src/Piping.Server.Core/Pipes/IPipe.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Piping.Server.Core.Pipes
+{
+    public interface IPipe
+    {
+        RequestKey Key { get; }
+        PipeStatus Status { get; }
+        bool IsRemovable { get; }
+        int RequestedReceiversCount { get; }
+        int ReceiversCount { get; }
+        event PipeStatusChangeEventHandler? OnStatusChanged;
+        event EventHandler? OnFinally;
+    }
+}
