+++ src/Piping.Server.Core/Pipes/IRecivePipe.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Piping.Server.Core.Pipes
+{
+    public interface IRecivePipe : IPipe
+    {
+        ValueTask ConnectionAsync(ICompletableStream CompletableStream, CancellationToken Token = default);
+    }
+}
