+++ src/Piping.Server.Core/Pipes/PipeStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace Piping.Server.Core.Pipes
+{
+    public enum PipeStatus : byte
+    {
+        Wait = 0,
+        Ready = 1,
+        ResponseStart = 2,
+        Canceled = 3,
+        Dispose = 4,
+    }
+}
