+++ src/Piping.Server.Core/Pipes/PipeStatusChangeEventHandler.cs
@@ -1,0 +1,4 @@
+﻿namespace Piping.Server.Core.Pipes
+{
+    public delegate void PipeStatusChangeEventHandler(object? sender, PipeStatusChangedArgs args);
+}
