+++ src/Piping.Server.Core/Pipes/PipeType.cs
@@ -1,0 +1,9 @@
+﻿namespace Piping.Server.Core.Pipes
+{
+    public enum PipeType : byte
+    {
+        None = 0,
+        Sender = 1,
+        Receiver = 2,
+    }
+}
