+++ test/Piping.Server.Core.Tests/Pipes/PipingProviderTests.cs
@@ -12,7 +12,7 @@ namespace Piping.Server.Core.Pipe.Tests
         {
             var services = new ServiceCollection();
             services.AddLogging();
-            services.AddTransient<IPipingProvider, PipingProvider>();
+            //services.AddTransient<IPipingProvider, PipingProvider>();
             var provider = services.BuildServiceProvider();
             var piping = provider.GetRequiredService<IPipingProvider>();
         }
